Complete support for C++ global and static constructors and deconstructors. The only caveat so far is that normal C programs will have "extra" sections, but these can be removed by ee-strip.  It's not a big deal seeing as how the linking step for ps2sdk was already hosed (see below).  The other thing I need to work out is how to get _start at the top of the program (I have an idea but it's untested).
I also had to change Rules.make to use -mno-crt0 instead of -nostartfiles
on the linker line.  When -nostartfiles was used, it prevented the _init
and _fini magic from being linked into the program.  This means that ps2sdk
never had a complete link as specified by the GNU toolchain.