Steps towards a v3.0 release. Using threads makes it snappier, request and response system rewritten to be cleaner code. Added some missing commands and changed the argument order to be more in line with the ps2link spec.
For the bad news, I may have broken win32 support but it was unintentional.
The readdir command is currently unimplemented, as is iopexcep. What the heck
is that?

The real bad news? Support for ps2netfs has been dropped. It never really
fit into what ps2client was supposed to be. I'll probably whip up a client
for it at some later point, but it really would be better served by a GUI.