WARNING: MAJOR GSKIT CHANGES AHEAD!
This is the first commit with the new accelerated render queue as well
as many other changes!

It is designed to be 100% backwards compatible, but you need to make sure to
cleanly recompile as some macros have changed.

The examples have been updated on how to "optimize" your gsKit app to take advantage
of the new render queue system.

There are still some misc bugs sitting around in the new code, if you find on in specific
please get as much info on it as you can and let me know so I can look into it.

Enjoy :)

(Also, once all known bugs in the new code are zapped, and it's performing well, I will
 make the long-awaited gsKit 0.3 release!)


PS: When used correctly, gsKit is now just over 3x faster than it used to be in regards to
    raw prim pushing power.