Several more optimizations!
From the changelog:

* More optimizations!
* Now using the scratchpad more intensively duing renderlist building.
* Packed the Element structure more tightly, and made it an aligned non-pointer array in the Queue.  
* NOTICE: Modified the GS_AXIS_ macros so that X, U, and V are the same value, so I only have to check on one in gsKit_scale. (Better performance)

Also, a quick performance update, we are up to 7900 triangles per frame (just shy of 400,000 tri/sec) now!

This is just under 4x as fast as gsKit 0.2! (and quite a bit faster than the last SVN commit even!)

There is a bug however, please read about it in the STATUS file.

Enjoy!